backend/requirements.txt
# Save as: apsara-beauty/backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# ML/AI
opencv-python==4.8.1.78
tensorflow==2.14.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
Pillow==10.1.0
mtcnn==0.1.1

# Cloud Services
boto3==1.29.7
google-api-python-client==2.108.0

# Task Queue
celery==5.3.4

# API Clients
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1